refactor: split config::Server and config::Upstream
