+++ src/config/mod.rs
@@ -3,6 +3,7 @@ pub use expr::*;
 pub use key_values::*;
 pub use server::*;
 pub use source::*;
+pub use upstream::*;
 mod config;
 mod expr;
 mod from_document;
@@ -13,6 +14,7 @@ mod n_plus_one;
 pub mod reader;
 mod server;
 mod source;
+mod upstream;
 
 fn is_default<T: Default + Eq>(val: &T) -> bool {
   *val == T::default()
